refactor(mongo): create main API function, and mongo type
created single function that handles native database connection, and the type for our responses
