--- src/pages/api/notion.ts
@@ -1,5 +1,0 @@
-import { Client } from '@notionhq/client';
-
-export const notion = new Client({ 
-    auth: process.env.NOTION_API_TOKEN 
-});
